+++ packages/blueberry/Widgets/UploadWidget/components/TakePhotoButton/constants.js
@@ -1,0 +1,25 @@
+export const initialFilesReferences = {
+	"id-front": "",
+	"id-back": "",
+	"residence-proof": "",
+	"income-proof": "",
+	"selfie": "",
+}
+
+export const initialLoadingState = {
+	front: false,
+	back: false,
+  }
+
+export const buttonTexts = {
+	initial: {
+		front: "Tirar foto da frente",
+		back: "Tirar foto do verso",
+	},
+	loading: "Enviando...",
+	finished:  {
+		front: "Foto da frente coletada",
+		back: "Foto do verso coletada",
+	},
+	error: "Houve um erro. Tente mais tarde."
+}