+++ packages/blueberry/utils/isClient.js
@@ -1,0 +1,8 @@
+const isClient = () =>
+  !!(
+    typeof window !== 'undefined' &&
+    window.document &&
+    window.document.createElement
+  )
+
+export default isClient