+++ packages/client/src/components/FireboltProvider/hook/index.ts
@@ -32,6 +32,8 @@ function useFireboltProvider({
     setIsFormLoading,
     formFlowHasBeenFinished,
     setFormFlowHasBeenFinished,
+    beforeProceedPayload,
+    setBeforeProceedPayload,
   } = useStates()
 
   const {
@@ -49,6 +51,7 @@ function useFireboltProvider({
     setStagedStep,
     lastVisitedStep,
     setLastVisitedStep,
+    clearRemoteFieldError,
   } = useData()
 
   useBrowserNavigation({
@@ -102,13 +105,16 @@ function useFireboltProvider({
     { extraRequestsMetaData = {} }: IRequestMetadata = {}
   ): Promise<void | Object> {
     setIsFormLoading(true)
+    setBeforeProceedPayload(stepFieldsPayload)
     const isLastStep = currentStep?.data?.slug === formflowMetadata?.lastStep
     return formEngine.current
       .nextStep(currentStep.data.slug, stepFieldsPayload, {
         extraRequestsMetaData,
       })
       .then((data) => {
-        if (isLastStep) {
+        const changedTrack = data?.step?.webhookResult?.['newTrackSlug']
+        
+        if (isLastStep && !changedTrack) {
           setFormEndPayload({
             webhookResult: data?.step?.webhookResult,
             capturedData: data?.capturedData,
@@ -189,7 +195,36 @@ function useFireboltProvider({
       setIsFormLoading(false)
 
       return { errors: invalidFields }
+    } else {
+      const unexpectedError = [
+        {
+          "slug": "unexpected_error",
+          "validationResults": [
+            {
+              "isValid": false,
+              "message": "Erro inesperado, por favor tente novamente",
+            },
+          ],
+        },
+      ]
+      setRemoteErrors(unexpectedError)
+      setIsFormLoading(false)
+
+      return { errors: unexpectedError }
     }
+  }
+
+  function addFieldRemoteError(fieldSlug: string, errorMessage: string) {
+    const newRemoteError = {
+      "slug": fieldSlug,
+      "validationResults": [
+        {
+          "isValid": false,
+          "message": errorMessage,
+        },
+      ],
+    }
+    setRemoteErrors([...remoteErrors, newRemoteError])
   }
 
   function uploadFile(file, fileName: string) {
@@ -218,6 +253,10 @@ function useFireboltProvider({
     getRequestsMetadata,
     uploadFile,
     clearSession,
+    clearRemoteFieldError,
+    beforeProceedPayload,
+    setBeforeProceedPayload,
+    addFieldRemoteError
   }
 }