+++ packages/client/src/types.ts
@@ -47,6 +47,7 @@ export interface IWizardHook {
   onConnectionError?(arg0?: object): void
   onFinishForm?(arg0?: object): void
   onBeforeChangeStep?(arg0?: Function, arg1?: IStepProps): void
+  onBeforeProceed?(sendingStep, formPayload): void
 }
 
 export interface IWizardComponent {
@@ -56,20 +57,33 @@ export interface IWizardComponent {
   onConnectionError?(arg0?: object): void
   onFinishForm?(arg0?: object): void
   onBeforeChangeStep?(arg0?: Function, arg1?: IStepProps): void
+  onBeforeProceed?(sendingStep, formPayload): void
 }
 
 export interface IUseFireboltForm {
   schema: Array<IStepConfigField>
   children?: Object[]
-  onChange?: React.ChangeEvent<HTMLInputElement>
+  onChange?: (formPayload: IFieldsObject) => void
   onSubmit?(): void
   theme?: Object
   autoFill?: IFieldsObject
   remoteErrors?: Array<IFieldsObject>
   onGoBack?(): void
   classes: Object
-  onFocusField?: Event
+  onFocusField?: (field: IStepConfigField, formPayload?: Object) => void
+  onBlurField?: (
+    field: IStepConfigField,
+    value: string,
+    formPayload?: Object
+  ) => void
+  onChangeField?: (
+    field: IStepConfigField,
+    values: { value: any; previousValue: any },
+    formPayload?: Object
+  ) => void
   addons?: IAddonsConfig
+  clearRemoteFieldError?: (fieldSlug: string) => void
+  orderFields?: Object[]
 }
 
 export interface IFormState {
@@ -109,11 +123,23 @@ export interface IFireboltForm {
   addons?: IAddonsConfig
   schema: Array<IStepConfigField>
   children?: Object[]
-  onChange?: React.ChangeEvent<HTMLInputElement>
+  onChange?: (formPayload: IFieldsObject) => void
   onSubmit?(): void
   theme?: Object
   autoFill?: IFieldsObject
   remoteErrors?: Array<IFieldsObject>
   onGoBack?(): void
-  onFocusField?: Event
+  onFocusField?: (field: IStepConfigField, formPayload?: Object) => void
+  onBlurField?: (
+    field: IStepConfigField,
+    value: string,
+    formPayload?: Object
+  ) => void
+  onChangeField?: (
+    field: IStepConfigField,
+    values: { value: any; previousValue: any },
+    formPayload?: Object
+  ) => void
+  clearRemoteFieldError?: (fieldSlug: string) => void
+  orderFields?: Object[]
 }