+++ packages/lab/src/app/pages/debug/Core.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState, useRef } from "react"
 // import { createFireboltForm, createPropsPreset } from "@iq-firebolt/client-core"
-import { createFireboltForm } from "@iq-firebolt/client-core/lib"
+import { createFireboltForm } from "@iq-firebolt/client-core"
 import * as S from "./styles.js"
 import transmorphersPropsPresets from "../../constants/transmorphers-props-presets"