+++ packages/validators/src/types.ts
@@ -28,3 +28,8 @@ export interface InvalidField {
   slug: string;
   validationResults: ValidationResult[];
 }
+
+export interface ValidateFBTStepResult {
+  isValid: boolean;
+  invalidFields: InvalidField[];
+}