+++ packages/validators/src/validators/noSpecialCharacters/noSpecialCharacters.test.ts
@@ -1,7 +1,7 @@
 import specialCharacters from './index';
 
 describe('specialCharacters validation', () => {
-  test.each(['geroge$', 'jorge`s'])(
+  test.each(['geroge$', 'jorge`s', '1234-5678'])(
     '%p should not pass in specialCharacters validator',
     (value) => {
       expect(specialCharacters.run(value).isValid).toBeFalsy();